+++ confmap/resolver.go
@@ -18,9 +18,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"reflect"
 	"regexp"
-	"strconv"
 	"strings"
 
 	"go.uber.org/multierr"
@@ -220,194 +218,4 @@ func (mr *Resolver) closeIfNeeded(ctx context.Context) error {
 	}
 	mr.closers = nil
 	return err
-}
-
-func (mr *Resolver) expandValueRecursively(ctx context.Context, value any) (any, error) {
-	for i := 0; i < 100; i++ {
-		val, changed, err := mr.expandValue(ctx, value)
-		if err != nil {
-			return nil, err
-		}
-		if !changed {
-			return val, nil
-		}
-		value = val
-	}
-	return nil, errTooManyRecursiveExpansions
-}
-
-func (mr *Resolver) expandValue(ctx context.Context, value any) (any, bool, error) {
-	switch v := value.(type) {
-	case string:
-		// No URIs to expand.
-		if !strings.Contains(v, "${") || !strings.Contains(v, "}") {
-			return value, false, nil
-		}
-
-		// Don't expand. Too many closing brackets.
-		if strings.Count(v, "}") > 100 {
-			return value, false, nil
-		}
-
-		URI, expand := mr.findURI(v)
-		// If the value is a single URI, then the return value can be anything.
-		// This is the case `foo: ${file:some_extra_config.yml}`.
-		if expand && URI == value {
-			return mr.expandStringURI(ctx, v)
-		}
-
-		// Embedded or nested URIs.
-		return mr.expandURIs(ctx, v)
-	case []any:
-		nslice := make([]any, 0, len(v))
-		nchanged := false
-		for _, vint := range v {
-			val, changed, err := mr.expandValue(ctx, vint)
-			if err != nil {
-				return nil, false, err
-			}
-			nslice = append(nslice, val)
-			nchanged = nchanged || changed
-		}
-		return nslice, nchanged, nil
-	case map[string]any:
-		nmap := map[string]any{}
-		nchanged := false
-		for mk, mv := range v {
-			val, changed, err := mr.expandValue(ctx, mv)
-			if err != nil {
-				return nil, false, err
-			}
-			nmap[mk] = val
-			nchanged = nchanged || changed
-		}
-		return nmap, nchanged, nil
-	}
-	return value, false, nil
-}
-
-// findURI will find the URI corresponding to the first closing bracket in input.
-// findURI is only called when input contains a }.
-func (mr *Resolver) findURI(input string) (string, bool) {
-	closeIndex := closeIndex(input)
-	openIndex := openIndex(input[:closeIndex+1])
-	// Should not expand because there is a missing ${.
-	if openIndex < 0 {
-		return "", false
-	}
-
-	URI := input[openIndex : closeIndex+1]
-
-	// Should not expand. This is expanded in the expandconverter.
-	if !strings.Contains(URI, ":") {
-		return "", false
-	}
-
-	return URI, true
-}
-
-func closeIndex(s string) int {
-	return strings.Index(s, "}")
-}
-
-func openIndex(s string) int {
-	return strings.LastIndex(s, "${")
-}
-
-func (mr *Resolver) expandURIs(ctx context.Context, input string) (string, bool, error) {
-	var err error
-	var changed bool
-
-	URI, expand := mr.findURI(input)
-	if expand {
-		input, changed, err = mr.expandableURI(ctx, input, URI)
-	}
-	// Check if other URIs are expandable.
-	if !expand {
-		input, changed, err = mr.nonExpandableURI(ctx, input)
-	}
-	return input, changed, err
-}
-
-func (mr *Resolver) expandableURI(ctx context.Context, input string, uri string) (string, bool, error) {
-	repl, changed, err := mr.expandURI(ctx, uri)
-	return strings.ReplaceAll(input, uri, repl), changed, err
-}
-
-// nonExpandableURI finds the next expandable URI in input and expands it.
-func (mr *Resolver) nonExpandableURI(ctx context.Context, input string) (string, bool, error) {
-	var err error
-	var expandedRemaining string
-	var changed bool
-
-	noExpand := input[:closeIndex(input)+1]
-	remaining := input[closeIndex(input)+1:]
-
-	if strings.Contains(remaining, "}") {
-		expandedRemaining, changed, err = mr.expandURIs(ctx, remaining)
-		return noExpand + expandedRemaining, changed, err
-	}
-	return input, changed, err
-}
-
-func (mr *Resolver) expandURI(ctx context.Context, value string) (string, bool, error) {
-	expanded, changed, err := mr.expandStringURI(ctx, value)
-	if err != nil {
-		return "", changed, err
-	}
-	val := reflect.ValueOf(expanded)
-	switch val.Kind() {
-	case reflect.String:
-		return val.String(), changed, err
-	case reflect.Int, reflect.Int32, reflect.Int64:
-		return strconv.FormatInt(val.Int(), 10), changed, err
-	case reflect.Float32, reflect.Float64:
-		return strconv.FormatFloat(val.Float(), 'f', -1, 64), changed, err
-	case reflect.Bool:
-		return strconv.FormatBool(val.Bool()), changed, err
-	default:
-		return value, changed, fmt.Errorf("expanding %v, expected string value type, got %T", value, expanded)
-	}
-}
-
-func (mr *Resolver) expandStringURI(ctx context.Context, uri string) (any, bool, error) {
-	lURI, err := newLocation(uri[2 : len(uri)-1])
-	if err != nil {
-		return nil, false, err
-	}
-	if strings.Contains(lURI.opaqueValue, "$") {
-		return nil, false, fmt.Errorf("the uri %q contains unsupported characters ('$')", lURI.asString())
-	}
-	ret, err := mr.retrieveValue(ctx, lURI)
-	if err != nil {
-		return nil, false, err
-	}
-	mr.closers = append(mr.closers, ret.Close)
-	val, err := ret.AsRaw()
-	return val, true, err
-}
-
-type location struct {
-	scheme      string
-	opaqueValue string
-}
-
-func (c location) asString() string {
-	return c.scheme + ":" + c.opaqueValue
-}
-
-func newLocation(uri string) (location, error) {
-	submatches := uriRegexp.FindStringSubmatch(uri)
-	if len(submatches) != 3 {
-		return location{}, fmt.Errorf("invalid uri: %q", uri)
-	}
-	return location{scheme: submatches[1], opaqueValue: submatches[2]}, nil
-}
-
-func (mr *Resolver) retrieveValue(ctx context.Context, uri location) (*Retrieved, error) {
-	p, ok := mr.providers[uri.scheme]
-	if !ok {
-		return nil, fmt.Errorf("scheme %q is not supported for uri %q", uri.scheme, uri.asString())
-	}
-	return p.Retrieve(ctx, uri.asString(), mr.onChange)
 }
